File structure adjusted, depth map texture creation moved inside the light class
