+++ RigelRenderer/include/publicUtility/Assert.hpp
@@ -1,15 +1,13 @@
 #pragma once
 
 #ifdef _DEBUG
-	#define ASSERT(cond)									   \
-		if (!cond) {										   \
-			std::cerr << "Assertion Error! " << "In file: " << \
-			__FILE__ << ", at line: " << __LINE__ << '\n';     \
-		}													   \
+	#define ASSERT(cond, message)											\
+		if (!cond) {														\
+			std::cerr << "Assertion Error! " << message << " In file: " <<	\
+			__FILE__ << ", at line: " << __LINE__ << '\n';					\
+			__debugbreak();													\
+		}																	\
 		else { }
 #else
-	#define ASSERT(cond)
-#endif 
-
-
-
+	#define ASSERT(cond, message)
+#endif
